Split up psinode into multiple source files
